requirements.txt
pip==21.2.4

numpy==1.21.5 

pandas==1.3.5

altair==4.2.0

streamlit==1.8.0

pydeck==0.7.1

click==8.0.4

psycopg2-binary==2.9.3
